requirements.txt
asgiref==3.8.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
distlib==0.3.5
Django==4.2.16
django-cors-headers==4.4.0
django-crispy-forms==1.14.0
django-currentuser==0.9.0
django-extensions==3.2.3
django-filter==25.1
django-tempus-dominus==5.1.2.17
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.0.2
Flask-JWT-Extended==4.4.4
fonttools==4.29.1
greenlet==1.1.3
h11==0.12.0
html5lib==1.1
idna==3.3
indian-cities==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mpmath==1.2.1
mysqlclient==2.1.1
numpy==1.21.4
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.3.4
Pillow==9.0.1
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.40
sqlparse==0.4.2
sympy==1.9
trio==0.19.0
trio-websocket==0.9.2
twilio==7.17.0
typing_extensions==4.12.2
tzdata==2022.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==1.26.7
user-agents==2.2.0
virtualenv==20.16.3
webencodings==0.5.1
Werkzeug==2.0.2
WooCommerce==3.0.0
wsproto==1.0.0
